pip_requirements.txt
flask-login==0.6.*
flask-bcrypt==1.0.*
cssutils==2.6.*